Got Player to Face Mouse
It occurred to me after I last made a commit that I had not made the player face the mouse while it moves, as would be necessary for our gun mechanics. This has been rectified, but the trello's audit log is in disarray from player movement getting moved around. Appropriately enough. Also commented the code.