Rudimentary Enemy Shooting and Player Death
Theoretically, the guards should shoot when they receive the Shoot order from beehave. Similarly, a bullet hitting the player should cause them to lose health and when health reaches zero it should move to the (basic- not final!!) death screen. However, as the AI is not yet set to shoot, I will hand testing of this over to Gabriel, who will handle the AI.